+++ airBnb.html
@@ -13,9 +13,8 @@
 
     <link rel="preconnect" href="https://fonts.gstatic.com">
     <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR&display=swap" rel="stylesheet">
-   
-
    </head>
+   
   <body>
      <!-- HEADER -->
      <div class="header2 hide_show">
@@ -62,11 +61,11 @@
  
                <div class="menu_center">
                   <label>
-                     <input type="radio" name="menu_choice"></input>
+                     <input type="radio" name="menu_choice" checked></input>
                      <span>숙소</span>
                   </label>
                   <label>
-                     <input type="radio" name="menu_choice" checked></input>
+                     <input type="radio" name="menu_choice"></input>
                      <span>체험</span>
                   </label>
                   <label>
