+++ class.ctrl.js
@@ -1,0 +1,11 @@
+export class ClassCtrl {
+   constructor(el) {
+      this.el = el;
+   }
+   show() {
+      this.el.classList.remove("hide_show");
+   }
+   hide() {
+      this.el.classList.add("hide_show");
+   }
+}
